should maybe work and not print anything
